+++ views/dashboard/footer/index.tsx
@@ -6,8 +6,8 @@ import { FormControl, SelectChangeEvent, TextField } from '@mui/material'
 import { enqueueSnackbar } from 'notistack'
 
 import { Add } from '@mui/icons-material'
-import ContactLinkCard from '@/components/dashboard/cards/contactLink'
-import IconSelectField from '@/components/form/fields/IconSelectField'
+import ContactLinkCard from '@/views/dashboard/footer/components/contactLink'
+import IconSelectField from '@/views/dashboard/footer/components/IconSelectField'
 import { CircleIconButton } from '@/components/buttons'
 import bannerUtils from '@/utils/banners'
 import { createContactLink } from '@/lib/db/cms/contact-links'
